Put hover inside a buildFHSUserEnv
Based on this PR: https://github.com/NixOS/nixpkgs/pull/80075

I exported the `go` and `gcc` packages so it was available inside the
FHS env and also fixed the X11 libraries, and this seems to do the
trick.

There is no more reason to wrapper with CPATH, and the only reason we
keep LD_LIBRARY_PATH is because hover is built outside the
buildFHSUserEnv so it wouldn't found the libraries otherwise.