Migrate all packages to unstable
The idea of this PR is to migrate all packages to a pinned commit in
nixpkgs-unstable. This brings some advantages:

- We ensure that every Nubanker uses the same derivations, including its
  dependencies
- Should avoid issues between multiple NixOS versions (i.e.: old stable
  vs. stable vs. unstable)
- Make sure our tooling is up-to-date

Of course, there is some disadvantages too:

- We either pin a commit and needs to bump it from time to time, or
  track master and possible lose control and risk some regressions

For now I think pinning should be ok. We could make a CI that would bump
the commit from time-to-time and asks for a review. Of course, this is
for a future PR.

BTW, since this commit uses unstable for everything, and `clojure-lsp`
is currently broken in unstable, this also backports a fix from
@ericdallo fork (until it is merged in upstream).